pyproject.toml
[tool.poetry]
name = "self-reply-bot"
version = "0.1.0"
description = ""
authors = ["scratchyone <scratchywon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.6.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.5"
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
